refactors auth controller and exception response filter
