+++ .env.development
@@ -1,10 +1,10 @@
 APP_ENV="development"
 DB_TYPE="postgres"
-DB_HOST="db"
-DB_PORT=5432
+DB_HOST="localhost"
+DB_PORT=8888
 DB_USERNAME='postgres'
-DB_PASSWORD='postgres'
-DB_NAME='conosco'
+DB_PASSWORD='docker'
+DB_NAME='conosco_dev'
 DB_ENTITIES_DIR='src/**/**/*.entity.ts'
 DB_MIGRATIONS_DIR='db/migrations/*.ts'
 DB_SUBS_DIR='src/subscriber/**/*.ts'
