+++ .gitignore
@@ -55,6 +55,7 @@ typings/
 .yarn-integrity
 
 # dotenv environment variables file
+
 .env.production
 .env
 dist
