+++ src/common/pipes/interceptors/exception.pipe.ts
@@ -5,41 +5,49 @@ import {
   Logger,
   HttpException,
   HttpStatus,
+  UseInterceptors,
 } from '@nestjs/common';
 
 import { EntityNotFoundError } from 'typeorm/error/EntityNotFoundError';
 import * as moment from 'moment';
+import { ResponseTransformInterceptor } from './response.pipe';
+import { ResponseDTO } from '../dto/response.dto';
 
-@Catch(HttpException, EntityNotFoundError, Error)
+@Catch()
+@UseInterceptors(ResponseTransformInterceptor)
 export class ExceptionInterceptor implements ExceptionFilter {
   catch(exception: any, host: ArgumentsHost) {
-    const ctx = host.switchToHttp();
-    const response = ctx.getResponse();
-    const request = ctx.getRequest();
-    let status: number;
+    let statusCode: number;
 
     if (exception instanceof EntityNotFoundError) {
-      status = HttpStatus.NOT_FOUND;
+      statusCode = HttpStatus.NOT_FOUND;
     } else {
-      status = exception.getStatus
+      statusCode = exception.getStatus
         ? exception.getStatus()
         : HttpStatus.INTERNAL_SERVER_ERROR;
     }
 
     const date = new Date();
     const timestamp = moment(date).format('YYYY-MM-DD');
+    const ctx = host.switchToHttp();
+    const response = ctx.getResponse();
+    const request = ctx.getRequest();
+    const error = true;
+    const data = null;
 
-    const errorResponse = {
-      statusCode: status,
+    let errorResponse = new ResponseDTO();
+
+    errorResponse = {
+      statusCode,
       timestamp,
-      error: true,
+      error,
       path: request.url,
       method: request.method,
       message:
-        status !== HttpStatus.INTERNAL_SERVER_ERROR
+        statusCode !== HttpStatus.INTERNAL_SERVER_ERROR
           ? exception.message.error || exception.message || null
           : 'Internal server error',
-      data: null,
+      data,
     };
 
     Logger.error(
@@ -48,6 +56,6 @@ export class ExceptionInterceptor implements ExceptionFilter {
       'ExceptionHttpFilter',
     );
 
-    response.status(status).json(errorResponse);
+    return response.status(statusCode).json(errorResponse);
   }
 }
